+++ lib/potential.js
@@ -1,11 +1,14 @@
 'use strict';
+
+//===== CORE PROMISES/A+ LIBRARY =====
+
 //===== Private =====
 
 // helpers for more readable code
 function invoke (fns) { fns.forEach(function (fn) { fn(); }); }
 function isFunction (maybe) { return typeof maybe === 'function'; }
 function isObjectOrFunction (maybe) {
-  return maybe && (typeof maybe).match(/function|object/);
+  return maybe && !!(typeof maybe).match(/function|object/);
 }
 
 // ancillary promise handler group constructor
@@ -37,11 +40,6 @@ $Promise.prototype.then = function then (onFulfilled, onRejected) {
   return newChain.downstream.promise;
 };
 
-// extra feature: sugar for adding rejection handlers
-$Promise.prototype.catch = function (onRejected) {
-  return this.then(null, onRejected);
-};
-
 $Promise.prototype.tryHandlers = function tryHandlers () {
   if (this.state === 'pending') return;
   var chain, handler, schedule = new Schedule();
@@ -67,8 +65,8 @@ function executes (handler, value, downstream) {
 // most of the complexity comes from "thenables", objets which may be promises
 function promiseResolutionProcedure (downstream, x) { // 2.3
   var promise = downstream.promise,
-      resolve = downstream.resolve.bind(downstream),
-      reject = downstream.reject.bind(downstream);
+      resolve = downstream.resolve,
+      reject = downstream.reject;
   if (promise === x) { // 2.3.1
     reject( new TypeError('handlers must not return current chain') );
   } else if ( x instanceof $Promise ) { // 2.3.2
@@ -127,7 +125,7 @@ Deferral.prototype.reject = function reject (reason) {
 function Potential (executor) {
   if (!isFunction(executor)) throw new TypeError('Potential takes a function');
   var deferral = Potential.defer();
-  executor( deferral.resolve.bind(deferral), deferral.reject.bind(deferral) );
+  executor( deferral.resolve, deferral.reject );
   return deferral.promise;
 }
 
@@ -149,6 +147,9 @@ Potential.rejected = function rejected (reason) {
   return new Deferral().reject(reason);
 };
 Potential.reject = Potential.rejected; // alias
+
+//----- Extend library with additional methods -----
+require('./extend')(Potential, $Promise);
 
 // It's alive!
 module.exports = Potential;
